# 18.5% performance improvement in MPPI