--- a/nav2_costmap_2d/include/nav2_costmap_2d/footprint_collision_checker.hpp
+++ b/nav2_costmap_2d/include/nav2_costmap_2d/footprint_collision_checker.hpp
@@ -51,11 +51,11 @@ public:
   /**
    * @brief Find the footprint cost in oriented footprint
    */
-  double footprintCost(const Footprint footprint);
+  double footprintCost(const Footprint & footprint);
   /**
    * @brief Find the footprint cost a a post with an unoriented footprint
    */
-  double footprintCostAtPose(double x, double y, double theta, const Footprint footprint);
+  double footprintCostAtPose(double x, double y, double theta, const Footprint & footprint);
   /**
    * @brief Get the cost for a line segment
    */