--- a/nav2_mppi_controller/include/nav2_mppi_controller/critics/cost_critic.hpp
+++ b/nav2_mppi_controller/include/nav2_mppi_controller/critics/cost_critic.hpp
@@ -56,7 +56,30 @@ protected:
     * @param theta theta of pose
     * @return bool if in collision
     */
-  bool inCollision(float cost, float x, float y, float theta);
+  inline bool inCollision(float cost, float x, float y, float theta)
+  {
+    // If consider_footprint_ check footprint scort for collision
+    float score_cost = cost;
+    if (consider_footprint_ &&
+      (cost >= possible_collision_cost_ || possible_collision_cost_ < 1.0f))
+    {
+      score_cost = static_cast<float>(collision_checker_.footprintCostAtPose(
+        static_cast<double>(x), static_cast<double>(y), static_cast<double>(theta),
+        costmap_ros_->getRobotFootprint())); // TODO float to double stuff
+        // TODO approximate with precomputed footprint orientations and translations?
+    }
+
+    switch (static_cast<unsigned char>(score_cost)) {
+      case (nav2_costmap_2d::LETHAL_OBSTACLE):
+        return true;
+      case (nav2_costmap_2d::INSCRIBED_INFLATED_OBSTACLE):
+        return consider_footprint_ ? false : true;
+      case (nav2_costmap_2d::NO_INFORMATION):
+        return is_tracking_unknown_ ? false : true;
+    }
+
+    return false;
+  }
 
   /**
     * @brief Find the min cost of the inflation decay function for which the robot MAY be
@@ -65,20 +88,57 @@ protected:
     * @return double circumscribed cost, any higher than this and need to do full footprint collision checking
     * since some element of the robot could be in collision
     */
-  float findCircumscribedCost(std::shared_ptr<nav2_costmap_2d::Costmap2DROS> costmap);
+  inline float findCircumscribedCost(std::shared_ptr<nav2_costmap_2d::Costmap2DROS> costmap);
 
-protected:
+  /**
+    * @brief An implementation of worldToMap fully using floats
+    * @param wx Float world X coord
+    * @param wy Float world Y coord
+    * @param mx unsigned int map X coord
+    * @param my unsigned into map Y coord
+    * @return if successsful
+    */
+  inline bool worldToMapFloat(float wx, float wy, unsigned int & mx, unsigned int & my) const
+  {
+    if (wx < origin_x_ || wy < origin_y_) {
+      return false;
+    }
+
+    mx = static_cast<unsigned int>((wx - origin_x_) / resolution_);
+    my = static_cast<unsigned int>((wy - origin_y_) / resolution_);
+
+    if (mx < size_x_ && my < size_y_) {
+      return true;
+    }
+    return false;
+  }
+
+  /**
+    * @brief An implementation of getIndex fully using floats
+    * @param mx unsigned int map X coord
+    * @param my unsigned into map Y coord
+    * @return Index
+    */
+  inline unsigned int getIndexFloat(unsigned int mx, unsigned int my) const
+  {
+    return static_cast<unsigned int>(my * size_x_ + mx);
+  }
+
   nav2_costmap_2d::FootprintCollisionChecker<nav2_costmap_2d::Costmap2D *>
   collision_checker_{nullptr};
   float possible_collision_cost_;
 
   bool consider_footprint_{true};
   bool is_tracking_unknown_{true};
-  float circumscribed_radius_{0};
-  float circumscribed_cost_{0};
-  float collision_cost_{0};
-  float critical_cost_{0};
+  float circumscribed_radius_{0.0f};
+  float circumscribed_cost_{0.0f};
+  float collision_cost_{0.0f};
+  float critical_cost_{0.0f};
   float weight_{0};
+  unsigned int trajectory_point_step_;
+
+  float origin_x_, origin_y_, resolution_;
+  unsigned int size_x_, size_y_;
 
   float near_goal_distance_;
   std::string inflation_layer_name_;