--- a/nav2_mppi_controller/src/critics/constraint_critic.cpp
+++ b/nav2_mppi_controller/src/critics/constraint_critic.cpp
@@ -55,12 +55,12 @@ void ConstraintCritic::score(CriticData & data)
           (std::move(
             xt::maximum(data.state.vx - max_vel_, 0.0f) +
             xt::maximum(min_vel_ - data.state.vx, 0.0f))) *
-        data.model_dt, {1}, immediate) * weight_, power_);
+          data.model_dt, {1}, immediate) * weight_, power_);
     } else {
       data.costs += xt::sum(
-          (std::move(
-            xt::maximum(data.state.vx - max_vel_, 0.0f) +
-            xt::maximum(min_vel_ - data.state.vx, 0.0f))) *
+        (std::move(
+          xt::maximum(data.state.vx - max_vel_, 0.0f) +
+          xt::maximum(min_vel_ - data.state.vx, 0.0f))) *
         data.model_dt, {1}, immediate) * weight_;
     }
     return;
@@ -77,12 +77,12 @@ void ConstraintCritic::score(CriticData & data)
           (std::move(
             xt::maximum(vel_total - max_vel_, 0.0f) +
             xt::maximum(min_vel_ - vel_total, 0.0f))) *
-        data.model_dt, {1}, immediate) * weight_, power_);
+          data.model_dt, {1}, immediate) * weight_, power_);
     } else {
       data.costs += xt::sum(
-          (std::move(
-            xt::maximum(vel_total - max_vel_, 0.0f) +
-            xt::maximum(min_vel_ - vel_total, 0.0f))) *
+        (std::move(
+          xt::maximum(vel_total - max_vel_, 0.0f) +
+          xt::maximum(min_vel_ - vel_total, 0.0f))) *
         data.model_dt, {1}, immediate) * weight_;
     }
     return;
@@ -101,12 +101,12 @@ void ConstraintCritic::score(CriticData & data)
           (std::move(
             xt::maximum(data.state.vx - max_vel_, 0.0f) +
             xt::maximum(min_vel_ - data.state.vx, 0.0f) + out_of_turning_rad_motion)) *
-        data.model_dt, {1}, immediate) * weight_, power_);
+          data.model_dt, {1}, immediate) * weight_, power_);
     } else {
       data.costs += xt::sum(
-          (std::move(
-            xt::maximum(data.state.vx - max_vel_, 0.0f) +
-            xt::maximum(min_vel_ - data.state.vx, 0.0f) + out_of_turning_rad_motion)) *
+        (std::move(
+          xt::maximum(data.state.vx - max_vel_, 0.0f) +
+          xt::maximum(min_vel_ - data.state.vx, 0.0f) + out_of_turning_rad_motion)) *
         data.model_dt, {1}, immediate) * weight_;
     }
     return;