From 12.7ms on average to 10.3ms on average for 10000 cycles 

I believe I also barely scratched the surface, the Cost, Path Align, and Path Angle critics can possibly  (definitely?) be improved more:
- Cost takes 2-3 ms 
- Path Align takes 1-2 ms
- Path Angle when triggers takes 0.4-1.1ms, which is mostly in `atan2` evaluation

That's not to even consider the trajectory roll outs and other areas for improvement in the `optimizer`. My aim is to halve again each of those 3 remaining critics. All other critics are reliably under 0.5ms at all times.