@SteveMacenski 2024-03-12
TODO apply to all critics

```
  if (power_ > 1u) {
    data.costs += xt::pow(std::move(cost) * weight_, power_);
  } else {
    data.costs += std::move(cost) * weight_;
  }
```

Makes 10x faster if we don't need to use power