@aosmw 2024-03-25
Hi Steve, 

Do you know about this "pitfall"?  https://xtensor.readthedocs.io/en/latest/pitfall.html#random-numbers-not-consistent
Using a random number function from xtensor actually returns a lazy generator. That means, accessing the same element of a random number generator does not give the same random number if called twice.

Is there a speed advantage if the xt::random::randn operations in NoiseGenerator::generateNoisedControls are wrapped in xt::eval?

regards
Mikew